Make PostProvider custom provider
Made a custom provider by creating a new file and taking all the context provider logic to the new file
